move translation keys
